feat: 평가 조회, 평가 수정 구현

+++ src/main/java/dingdong/dingdong/dto/rating/RatingResponseDto.java
@@ -12,11 +12,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class RatingResponseDto {
 
-    private int good;
+    private Long good;
 
-    private int bad;
+    private Long bad;
 
-    private int total;
+    private Long total;
 
     public static RatingResponseDto from(Profile profile) {
         return RatingResponseDto.builder()
